Add on_event to python subscriber
