+++ src/daft-context/src/subscribers/python.rs
@@ -4,7 +4,10 @@ use async_trait::async_trait;
 use common_error::DaftResult;
 use common_metrics::{QueryID, QueryPlan, Stats};
 use daft_micropartition::{MicroPartitionRef, python::PyMicroPartition};
-use pyo3::{IntoPyObject, Py, PyAny, Python, intern};
+use pyo3::{
+    Bound, IntoPyObject, Py, PyAny, PyResult, Python, intern,
+    types::{PyAnyMethods, PyModule},
+};
 
 use crate::{
     python::{PyQueryMetadata, PyQueryResult},
@@ -17,55 +20,6 @@ use crate::{
 /// Wrapper around a Python object that implements the Subscriber trait
 #[derive(Debug)]
 pub struct PySubscriberWrapper(pub(crate) Py<PyAny>);
-
-impl PySubscriberWrapper {
-    fn py_operator_start(&self, event: Arc<OperatorStartEvent>) -> DaftResult<()> {
-        Python::attach(|py| {
-            self.0.call_method1(
-                py,
-                intern!(py, "on_operator_start"),
-                (event.header.query_id.to_string(), event.operator.node_id),
-            )?;
-            Ok(())
-        })
-    }
-
-    fn py_operator_end(&self, event: Arc<OperatorEndEvent>) -> DaftResult<()> {
-        Python::attach(|py| {
-            self.0.call_method1(
-                py,
-                intern!(py, "on_operator_end"),
-                (event.header.query_id.to_string(), event.operator.node_id),
-            )?;
-            Ok(())
-        })
-    }
-
-    fn py_stats(&self, event: Arc<StatsEvent>) -> DaftResult<()> {
-        Python::attach(|py| {
-            let stats_map = event
-                .stats
-                .iter()
-                .map(|(node_id, stats)| {
-                    let stat_map = stats
-                        .iter()
-                        .map(|(name, stat)| {
-                            (name.to_string(), stat.clone().into_py_contents(py).unwrap())
-                        })
-                        .collect::<HashMap<_, _>>();
-                    (*node_id, stat_map)
-                })
-                .collect::<HashMap<_, _>>();
-            let py_stats = stats_map.into_pyobject(py)?;
-            self.0.call_method1(
-                py,
-                intern!(py, "on_stats"),
-                (event.header.query_id.to_string(), py_stats),
-            )?;
-            Ok(())
-        })
-    }
-}
 
 #[async_trait]
 impl Subscriber for PySubscriberWrapper {
@@ -214,11 +168,71 @@ impl Subscriber for PySubscriberWrapper {
     }
 
     async fn on_event(&self, event: Event) -> DaftResult<()> {
-        match event {
-            Event::Stats(e) => self.py_stats(e)?,
-            Event::OperatorStart(e) => self.py_operator_start(e)?,
-            Event::OperatorEnd(e) => self.py_operator_end(e)?,
-        }
-        Ok(())
+        Python::attach(|py| {
+            let py_event = build_py_event(py, event)?;
+            self.0
+                .call_method1(py, intern!(py, "on_event"), (py_event,))?;
+            Ok(())
+        })
+    }
+}
+
+// Build Python subscriber event objects from Rust execution events.
+
+fn events_module(py: Python<'_>) -> PyResult<Bound<'_, PyModule>> {
+    py.import("daft.subscribers.events")
+}
+
+fn event_class<'py>(py: Python<'py>, name: &str) -> PyResult<Bound<'py, PyAny>> {
+    events_module(py)?.getattr(name)
+}
+
+fn build_operator_started(py: Python<'_>, event: Arc<OperatorStartEvent>) -> PyResult<Py<PyAny>> {
+    event_class(py, "OperatorStarted")?
+        .call1((
+            event.header.query_id.to_string(),
+            event.operator.node_id,
+            event.operator.name.as_ref(),
+        ))
+        .map(Into::into)
+}
+
+fn build_operator_finished(py: Python<'_>, event: Arc<OperatorEndEvent>) -> PyResult<Py<PyAny>> {
+    event_class(py, "OperatorFinished")?
+        .call1((
+            event.header.query_id.to_string(),
+            event.operator.node_id,
+            event.operator.name.as_ref(),
+        ))
+        .map(Into::into)
+}
+
+fn build_py_stats<'py>(py: Python<'py>, stats: &[(NodeID, Stats)]) -> PyResult<Bound<'py, PyAny>> {
+    let stats_map = stats
+        .iter()
+        .map(|(node_id, stats)| {
+            let stat_map = stats
+                .iter()
+                .map(|(name, stat)| Ok((name.to_string(), stat.clone().into_py_contents(py)?)))
+                .collect::<PyResult<HashMap<_, _>>>()?;
+            Ok((*node_id, stat_map))
+        })
+        .collect::<PyResult<HashMap<_, _>>>()?;
+
+    stats_map.into_pyobject(py).map(|obj| obj.into_any())
+}
+
+fn build_stats(py: Python<'_>, event: Arc<StatsEvent>) -> PyResult<Py<PyAny>> {
+    let py_stats = build_py_stats(py, event.stats.as_ref())?;
+    event_class(py, "Stats")?
+        .call1((event.header.query_id.to_string(), py_stats))
+        .map(Into::into)
+}
+
+fn build_py_event(py: Python<'_>, event: Event) -> PyResult<Py<PyAny>> {
+    match event {
+        Event::OperatorStart(event) => build_operator_started(py, event),
+        Event::OperatorEnd(event) => build_operator_finished(py, event),
+        Event::Stats(event) => build_stats(py, event),
     }
 }
